connection-manager: numberOfConnections
For connection manager which allows to include inbound connections we
need a method to count how many connections are currently included (for
rate limiting).  We count all connection, weather inbound or outbound.
In future we could count only inbound connections, but this will require
to know the state of mux (to know when on-demand responders actaully
started).